# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Swagger Demo",
        "description": "Demonstration of Swagger with Laravel",
        "contact": {
            "email": "hello@example.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1"
        }
    ],
    "paths": {
        "/auth/sign-up": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Sign up",
                "description": "Sign up",
                "operationId": "e90fee7993f86a751279c22ea41f9210",
                "requestBody": {
                    "description": "User info",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Passw0rd"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful response"
                    },
                    "422": {
                        "description": "Validation exception"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/auth/sign-in": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Sign in",
                "description": "Sign in",
                "operationId": "d4116d4e23a2d276cc449283feb4babd",
                "requestBody": {
                    "description": "User info",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Passw0rd"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "description": "Validation exception"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "Logout",
                "operationId": "a05022a34eef599e04fb0a5254375b84",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "description": "Validation exception"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/confirm-email/{confirmationToken}": {
            "get": {
                "summary": "Confirm email",
                "operationId": "fcf458b8b03117b0957acbd2ab062496",
                "parameters": [
                    {
                        "name": "confirmationToken",
                        "in": "path",
                        "description": "Confirm token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1234567890"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email confirmed successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid confirm token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/send-confirm-email": {
            "post": {
                "summary": "Send confirm email",
                "operationId": "b1d6db29f9dcb65571091e929e1d6cb3",
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email sent successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid confirm token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "parameters": {
            "service": {
                "name": "service_auth",
                "in": "path",
                "description": "Service of auth",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
            "service_email": {
                "name": "service_email",
                "in": "path",
                "description": "Service of email",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
            "service_token": {
                "name": "service_token",
                "in": "path",
                "description": "Service of token",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
            "service_email_confirmation": {
                "name": "service_email_confirmation",
                "in": "path",
                "description": "Service of confirmation",
                "required": true,
                "schema": {
                    "type": "string"
                }
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "API Endpoints of Auth"
        }
    ]
}